Flowsource-templates/python-gcp-ci-template/skeleton/requirements.txt
fastapi==0.115.2
hyx==0.0.2
pydantic==2.8.2
pydantic_settings==2.4.0
pytest==9.0.2
python-dotenv==1.0.1
pytest-asyncio==1.3.0
uvicorn==0.22.0
